+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* New `test_server()` to launch a local server, making tests and examples 
+  more robust.
+
 * New `throttle_status()` to make it a little easier to verify what's happening
   with throttling.